Updated event listener clases such that they can now consume events, and can have a priority. Also fixed some minor bugs/typos/changed path for include path for some modules

+++ src/zore/graphics/RenderEngine.cpp
@@ -32,7 +32,7 @@ namespace zore {
 	}
 
 	void RenderEngine::SetIndexType(IndexType type) {
-		static const uint64_t IndexTypeToGLIndexType[] = { GL_UNSIGNED_SHORT, GL_UNSIGNED_INT };
+		static const uint32_t IndexTypeToGLIndexType[] = { GL_UNSIGNED_SHORT, GL_UNSIGNED_INT };
 		static const uint64_t IndexTypeToIndexSize[] = { sizeof(uint16_t), sizeof(uint32_t) };
 		indexType = IndexTypeToGLIndexType[static_cast<uint32_t>(type)];
 		indexSize = IndexTypeToIndexSize[static_cast<uint32_t>(type)];
